update/clean: removed unused line in update, added to delete to return false

+++ Projects/Project0/src/main/java/dev/wan/services/ClientServiceImpl.java
@@ -34,13 +34,15 @@ public class ClientServiceImpl implements ClientService{
 
     @Override
     public Client updateClient(Client client) {
-//        Client oldClient = this.clientDao.getClientById(client.getId());
         this.clientDao.updateClient(client);
         return client;
     }
 
     @Override
     public boolean deleteClientById(int id) {
+        if(this.clientDao.getClientById(id) == null){
+            return false;
+        }
         return this.clientDao.deleteClientById(id);
     }
 }
